# api/backend/src/lambda/swagger.yaml
openapi: 3.0.3
info:
  title: FindNest API - Optimized
  description: |
    API tối ưu cho hệ thống tìm kiếm phòng trọ thông minh FindNest

    ## 3 Loại người dùng:

    ### 1. Người dùng (Users) - Tìm phòng:
    - Đăng nhập: Phone + OTP (tự động tạo tài khoản)
    - Quyền: Xem danh sách phòng, tìm kiếm, xem chi tiết, lưu yêu thích
    - Liên hệ: Gửi yêu cầu hỗ trợ, góp ý, báo lỗi

    ### 2. Chủ trọ (Landlords) - Đăng tin:
    - Đăng nhập: Phone + OTP, sau đó admin cấp quyền landlord
    - Quyền: Tạo, sửa, xóa tin đăng phòng trọ, upload hình ảnh

    ### 3. Quản trị viên (Admins) - Quản lý:
    - Đăng nhập: Username + Password
    - Quyền: Quản lý toàn hệ thống, xử lý yêu cầu hỗ trợ

    ## Tính năng chính:
    - Tìm kiếm thông minh bằng AI với ngôn ngữ tự nhiên
    - Giao diện bản đồ tương tác (Google Maps style)
    - Lọc theo vị trí và khoảng cách
    - Cá nhân hóa dựa trên favorites và profile
    - Hệ thống thông báo thông minh
    - Quản lý quyền người dùng

  version: 3.0.0
  contact:
    name: FindNest Team
    email: admin@findnest.com
servers:
  - url: https://k0qsj9xb67.execute-api.us-east-1.amazonaws.com/prod
    description: Production server
  - url: http://localhost:3001
    description: Local development server
...
paths:
...
  /listings/{listingId}:
    get:
      tags: [Listings]
      summary: Lấy chi tiết phòng trọ
      description: |
        Lấy thông tin chi tiết của một phòng trọ

        **Public endpoint** - Không cần đăng nhập
      parameters:
        - name: listingId
          in: path
          required: true
          description: ID của tin đăng
          schema:
            type: string
            example: "12345678-1234-1234-1234-123456789012"
      responses:
        "200":
          description: Chi tiết phòng trọ
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
        "404":
          description: Không tìm thấy tin đăng

    put:
      tags: [Listings]
      summary: Cập nhật tin đăng
      description: |
        Cập nhật thông tin tin đăng phòng trọ

        **Yêu cầu quyền:** Chỉ chủ sở hữu tin đăng hoặc Admins
      security:
        - bearerAuth: []
      parameters:
        - name: listingId
          in: path
          required: true
          description: ID của tin đăng
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateListingRequest"
      responses:
        "200":
          description: Cập nhật thành công
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
        "400":
          description: Dữ liệu không hợp lệ
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: Không có quyền cập nhật tin đăng này
        "404":
          description: Không tìm thấy tin đăng

    delete:
      tags: [Listings]
      summary: Xóa tin đăng
      description: |
        Xóa tin đăng phòng trọ

        **Yêu cầu quyền:** Chỉ chủ sở hữu tin đăng hoặc Admins
      security:
        - bearerAuth: []
      parameters:
        - name: listingId
          in: path
          required: true
          description: ID của tin đăng
          schema:
            type: string
      responses:
        "204":
          description: Xóa thành công (không có nội dung trả về)
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: Không có quyền xóa tin đăng này
        "404":
          description: Không tìm thấy tin đăng

  # ============================================================================
  # PROFILE ENDPOINTS
  # ============================================================================
...
  /admin/users:
    get:
      tags: [Admin]
      summary: Lấy danh sách tất cả users
      description: |
        Lấy danh sách tất cả users trong hệ thống với phân trang và lọc

        **Yêu cầu quyền:** Chỉ Admins
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          description: Số trang (bắt đầu từ 1)
          schema:
            type: integer
            default: 1
            example: 1
        - name: limit
          in: query
          description: Số lượng users mỗi trang
          schema:
            type: integer
            default: 20
            example: 20
        - name: userType
          in: query
          description: Lọc theo loại user
          schema:
            type: string
            enum: [user, landlord, admin]
            example: "user"
        - name: status
          in: query
          description: Lọc theo trạng thái
          schema:
            type: string
            enum: [ENABLED, DISABLED]
            example: "ENABLED"
      responses:
        "200":
          description: Danh sách users
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  users:
                    type: array
                    items:
                      oneOf:
                        - $ref: "#/components/schemas/UserProfile"
                        - $ref: "#/components/schemas/LandlordProfile"
                        - $ref: "#/components/schemas/AdminProfile"
                  pagination:
                    type: object
                    properties:
                      page:
                        type: integer
                        example: 1
                      limit:
                        type: integer
                        example: 20
                      total:
                        type: integer
                        example: 150
                      totalPages:
                        type: integer
                        example: 8
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/stats:
    get:
      tags: [Admin]
      summary: Lấy thống kê hệ thống
      description: |
        Lấy thống kê tổng số người dùng, phòng trọ, và phòng đang hoạt động

        **Yêu cầu quyền:** Chỉ Admins
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Thống kê hệ thống
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  stats:
                    type: object
                    properties:
                      totalUsers:
                        type: integer
                        description: Tổng số người dùng
                        example: 1250
                      totalListings:
                        type: integer
                        description: Tổng số phòng trọ
                        example: 450
                      activeListings:
                        type: integer
                        description: Số phòng đang hoạt động (chưa hết hạn)
                        example: 420
                      usersByType:
                        type: object
                        description: Phân loại người dùng theo role
                        properties:
                          user:
                            type: integer
                            example: 1000
                          landlord:
                            type: integer
                            example: 200
                          admin:
                            type: integer
                            example: 50
                      systemHealth:
                        type: string
                        description: Trạng thái hệ thống
                        example: "OK"
                      lastUpdated:
                        type: string
                        format: date-time
                        description: Thời gian cập nhật thống kê
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/listings/{listingId}:
    delete:
      tags: [Admin]
      summary: Xóa bài đăng vi phạm
      description: |
        Xóa các bài đăng vi phạm quy định

        **Yêu cầu quyền:** Chỉ Admins
      security:
        - bearerAuth: []
      parameters:
        - name: listingId
          in: path
          required: true
          description: ID của tin đăng
          schema:
            type: string
            example: "12345678-1234-1234-1234-123456789012"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason:
                  type: string
                  example: "Nội dung không phù hợp, hình ảnh không liên quan"
                  description: "Lý do xóa bài đăng"
      responses:
        "200":
          description: Xóa bài đăng thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Listing deleted successfully"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Không tìm thấy tin đăng
  /admin/users/{userId}/role:
    put:
      tags: [Admin]
      summary: Cập nhật quyền người dùng
      description: |
        Quản lý quyền (role) của người dùng

        **Yêu cầu quyền:** Chỉ Admins
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          description: ID của user
          schema:
            type: string
            example: "12345678-1234-1234-1234-123456789012"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role]
              properties:
                role:
                  type: string
                  enum: [user, landlord, admin]
                  example: "landlord"
                  description: "Quyền mới của người dùng"
      responses:
        "200":
          description: Cập nhật quyền thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "User role updated to landlord"
                  user:
                    type: object
                    description: "Thông tin user đã cập nhật"
        "400":
          description: Quyền không hợp lệ
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Không tìm thấy user
  /admin/users/{userId}/status:
    put:
      tags: [Admin]
      summary: Cập nhật trạng thái user
      description: |
        Khóa hoặc mở khóa tài khoản user

        **Yêu cầu quyền:** Chỉ Admins
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          description: ID của user
          schema:
            type: string
            example: "12345678-1234-1234-1234-123456789012"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [ENABLED, DISABLED]
                  example: "DISABLED"
                  description: "ENABLED = mở khóa, DISABLED = khóa tài khoản"
      responses:
        "200":
          description: Cập nhật trạng thái thành công
        "400":
          description: Trạng thái không hợp lệ
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Không tìm thấy user
  /admin/system/cleanup-inactive-users:
    post:
      tags: [Admin]
      summary: Xóa tài khoản không hoạt động
      description: |
        Xóa các tài khoản user không hoạt động quá 6 tháng (US-010)

        **Yêu cầu quyền:** Chỉ Admins
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Xóa thành công
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/system/refresh-area-data:
    post:
      tags: [Admin]
      summary: Làm mới dữ liệu khu vực
      description: |
        Làm mới thông tin ngữ cảnh khu vực cho các listing (US-040)

        **Yêu cầu quyền:** Chỉ Admins
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Làm mới thành công
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/notifications/send-weekly-digest:
    post:
      tags: [Admin]
      summary: Gửi bản tin hàng tuần
      description: |
        Gửi bản tin tổng hợp hàng tuần cho người dùng (US-050)

        **Yêu cầu quyền:** Chỉ Admins
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Gửi thành công
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/analytics/popular-areas:
    get:
      tags: [Analytics]
      summary: Thống kê khu vực phổ biến
      description: |
        Lấy thống kê các khu vực được tìm kiếm nhiều nhất (US-045)

        **Yêu cầu quyền:** Chỉ Admins
      security:
        - bearerAuth: []
      parameters:
        - name: days
          in: query
          description: Số ngày thống kê
          schema:
            type: integer
            default: 7
      responses:
        "200":
          description: Thống kê khu vực
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/analytics/users-by-area:
    get:
      tags: [Analytics]
      summary: Thống kê người dùng theo khu vực
      description: |
        Thống kê số lượng người dùng theo khu vực (US-060)

        **Yêu cầu quyền:** Chỉ Admins
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Thống kê người dùng
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "BadRequest"
        message:
          type: string
          example: "Invalid input data"
...
    UserProfile:
      type: object
      properties:
        userId:
          type: string
          example: "12345678-1234-1234-1234-123456789012"
        userType:
          type: string
          example: "user"
        phoneNumber:
          type: string
          example: "+84987654321"
        fullName:
          type: string
          example: "Nguyen Van A"
        createdAt:
          type: string
          format: date-time
        lastLoginAt:
          type: string
          format: date-time
    LandlordProfile:
      allOf:
        - $ref: "#/components/schemas/UserProfile"
        - type: object
          properties:
            userType:
              type: string
              example: "landlord"
            businessName:
              type: string
              example: "Nha Tro ABC"
            businessAddress:
              type: string
              example: "123 Nguyen Hue, Q1, TPHCM"
    AdminProfile:
      allOf:
        - $ref: "#/components/schemas/UserProfile"
        - type: object
          properties:
            userType:
              type: string
              example: "admin"
            email:
              type: string
              example: "admin@example.com"
    Listing:
      type: object
      properties:
        listingId:
          type: string
          example: "12345678-1234-1234-1234-123456789012"
        title:
          type: string
          example: "Phòng trọ đẹp gần trung tâm"
        description:
          type: string
          example: "Phòng trọ sạch sẽ, đầy đủ tiện nghi"
        price:
          type: number
          example: 3500000
        area:
          type: number
          example: 25
        address:
          type: object
          properties:
            street:
              type: string
              example: "123 Nguyen Hue"
            ward:
              type: string
              example: "Phuong Ben Nghe"
            district:
              type: string
              example: "Quan 1"
            city:
              type: string
              example: "TP.HCM"
        location:
          type: object
          properties:
            latitude:
              type: number
              example: 10.7769
            longitude:
              type: number
              example: 106.7009
        images:
          type: array
          items:
            type: string
            example: "https://bucket.s3.region.amazonaws.com/listings/userId/uuid-room1.jpg"
        amenities:
          type: array
          items:
            type: string
          example: ["WiFi", "Điều hòa", "Tủ lạnh"]
        ownerId:
          type: string
          example: "12345678-1234-1234-1234-123456789012"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    UpdateListingRequest:
      type: object
      properties:
        title:
          type: string
          example: "Phòng trọ đẹp gần trung tâm"
        description:
          type: string
          example: "Phòng trọ sạch sẽ, đầy đủ tiện nghi"
        price:
          type: number
          example: 3500000
        area:
          type: number
          example: 25
        images:
          type: array
          items:
            type: string
          example:
            [
              "https://bucket.s3.region.amazonaws.com/listings/userId/uuid-room1.jpg",
            ]
        amenities:
          type: array
          items:
            type: string
          example: ["WiFi", "Điều hòa", "Tủ lạnh"]
...
  responses:
...
    Unauthorized:
      description: Không có quyền truy cập - cần đăng nhập
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Không có quyền thực hiện hành động này
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...